#202: migrate BikeSharingDemand F# sample to v0.9

+++ samples/fsharp/getting-started/Regression_BikeSharingDemand/BikeSharingDemand/BikeSharingDemandConsoleApp/ModelScoringTester.fs
@@ -40,5 +40,5 @@ let visualizeSomePredictions testDataLocation modelScorer numberOfPredictions =
             modelScorer 
             |> Common.ModelScorer.predictSingle testData.[i]
 
-        Common.ConsoleHelper.printRegressionPredictionVersusObserved (int prediction.PredictedCount) (testData.[i].Count.ToString())
+        Common.ConsoleHelper.printRegressionPredictionVersusObserved (float prediction.PredictedCount) (testData.[i].Count.ToString())
